Initial import of the re-implementation of portmaster in LUA
This commit is pre-alpha quality, but options processing and loading
of rc files is already working, a usage message with all (to be)
supported options can be requested.

Options -a and -l are implemented, with -a only generating a list of
operations to be performed. Both are used to test and optimize low level
functions required to implement more complex functionality.

The time required for the --list operation is 2.5 seconds on my system
with more than 2000 installed ports vs. 13.5 seconds for portmaster.

Scanning all ports for updates takes 65 seconds (portmaster stops
due to installed ports that have been deleted in the ports tree, but
last time I measured the run time it needed some 250 to 400 seconds).