Add simple lock module
Lock structures have to be allocated with new() and can then acquired
in a blocking or non-blocking way with acquire(). A release() call is
required to let blocked competing tasks continue execution.

No attempt is made to detect or prevent deadlocks. If multiple items
need to be locked then each task should do so in the same order (e.g.
lexicographically sorted) to avoid deadlocks.